refactor: group definition by type
