+++ typegraph/core/src/typedef/mod.rs
@@ -1,0 +1,15 @@
+// Copyright Metatype OÜ, licensed under the Mozilla Public License Version 2.0.
+// SPDX-License-Identifier: MPL-2.0
+
+pub mod array;
+pub mod boolean;
+pub mod either;
+pub mod func;
+pub mod integer;
+pub mod number;
+pub mod optional;
+pub mod policy;
+pub mod proxy;
+pub mod string;
+pub mod struct_;
+pub mod union;
