+++ typegraph/core/src/types.rs
@@ -9,8 +9,8 @@ use crate::errors::Result;
 use crate::global_store::{with_store, Store};
 use crate::typegraph::TypegraphContext;
 use crate::wit::core::{
-    PolicySpec, TypeArray, TypeBase, TypeEither, TypeFunc, TypeId, TypeInteger, TypeNumber,
-    TypeOptional, TypePolicy, TypeProxy, TypeString, TypeStruct, TypeUnion,
+    TypeArray, TypeBase, TypeEither, TypeFunc, TypeId, TypeInteger, TypeNumber, TypeOptional,
+    TypePolicy, TypeProxy, TypeString, TypeStruct, TypeUnion,
 };
 
 pub trait TypeData {
@@ -128,221 +128,5 @@ where
 
     fn get_base(&self) -> Option<&TypeBase> {
         None
-    }
-}
-
-impl TypeData for TypeProxy {
-    fn get_display_params_into(&self, params: &mut Vec<String>) {
-        params.push(format!("proxy_name='{}'", self.name));
-    }
-
-    fn variant_name(&self) -> String {
-        "proxy".to_string()
-    }
-}
-
-impl WrapperTypeData for TypeProxy {
-    fn get_wrapped_type<'a>(&self, store: &'a Store) -> Option<&'a Type> {
-        store
-            .get_type_by_name(&self.name)
-            .map(|id| store.get_type(id).unwrap())
-    }
-}
-
-impl TypeData for TypeInteger {
-    fn get_display_params_into(&self, params: &mut Vec<String>) {
-        if let Some(min) = self.min {
-            params.push(format!("min={}", min));
-        }
-        if let Some(max) = self.max {
-            params.push(format!("max={}", max));
-        }
-        if let Some(exclusive_minimum) = self.exclusive_minimum {
-            params.push(format!("exclusiveMin={}", exclusive_minimum));
-        }
-        if let Some(exclusive_maximum) = self.exclusive_maximum {
-            params.push(format!("exclusiveMax={}", exclusive_maximum));
-        }
-        if let Some(multiple_of) = self.multiple_of {
-            params.push(format!("multipleOf={}", multiple_of));
-        }
-    }
-
-    fn variant_name(&self) -> String {
-        "integer".to_string()
-    }
-}
-
-impl TypeData for TypeNumber {
-    fn get_display_params_into(&self, params: &mut Vec<String>) {
-        if let Some(min) = self.min {
-            params.push(format!("min={}", min));
-        }
-        if let Some(max) = self.max {
-            params.push(format!("max={}", max));
-        }
-        if let Some(exclusive_minimum) = self.exclusive_minimum {
-            params.push(format!("exclusiveMin={}", exclusive_minimum));
-        }
-        if let Some(exclusive_maximum) = self.exclusive_maximum {
-            params.push(format!("exclusiveMax={}", exclusive_maximum));
-        }
-        if let Some(multiple_of) = self.multiple_of {
-            params.push(format!("multipleOf={}", multiple_of));
-        }
-    }
-
-    fn variant_name(&self) -> String {
-        "number".to_string()
-    }
-}
-
-impl TypeData for TypeBoolean {
-    fn get_display_params_into(&self, _params: &mut Vec<String>) {}
-
-    fn variant_name(&self) -> String {
-        "boolean".to_string()
-    }
-}
-
-impl TypeData for TypeString {
-    fn get_display_params_into(&self, params: &mut Vec<String>) {
-        if let Some(min) = self.min {
-            params.push(format!("min={}", min));
-        }
-        if let Some(max) = self.max {
-            params.push(format!("max={}", max));
-        }
-        if let Some(pattern) = self.pattern.to_owned() {
-            params.push(format!("pattern={}", pattern));
-        }
-        if let Some(format) = self.format.to_owned() {
-            params.push(format!("format={}", format));
-        }
-    }
-
-    fn variant_name(&self) -> String {
-        "string".to_string()
-    }
-}
-
-impl TypeData for TypeArray {
-    fn get_display_params_into(&self, params: &mut Vec<String>) {
-        params.push(format!("items={}", self.of));
-        if let Some(min) = self.min {
-            params.push(format!("minItems={}", min));
-        }
-        if let Some(max) = self.max {
-            params.push(format!("maxItems={}", max));
-        }
-        if let Some(unique) = self.unique_items {
-            params.push(format!("uniqueItems={}", unique));
-        }
-    }
-
-    fn variant_name(&self) -> String {
-        "array".to_string()
-    }
-}
-
-impl TypeData for TypeOptional {
-    fn get_display_params_into(&self, params: &mut Vec<String>) {
-        params.push(format!("item={}", self.of));
-        if let Some(default) = self.default_item.clone() {
-            params.push(format!("defaultItem={}", default));
-        }
-    }
-
-    fn variant_name(&self) -> String {
-        "optional".to_string()
-    }
-}
-
-impl TypeData for TypeUnion {
-    fn get_display_params_into(&self, params: &mut Vec<String>) {
-        for (i, tpe_id) in self.variants.iter().enumerate() {
-            params.push(format!("[v{}] => #{}", i, tpe_id));
-        }
-    }
-
-    fn variant_name(&self) -> String {
-        "union".to_string()
-    }
-}
-
-impl TypeData for TypeEither {
-    fn get_display_params_into(&self, params: &mut Vec<String>) {
-        for (i, tpe_id) in self.variants.iter().enumerate() {
-            params.push(format!("[v{}] => #{}", i, tpe_id));
-        }
-    }
-
-    fn variant_name(&self) -> String {
-        "either".to_string()
-    }
-}
-
-impl TypeData for TypeStruct {
-    fn get_display_params_into(&self, params: &mut Vec<String>) {
-        for (name, tpe_id) in self.props.iter() {
-            params.push(format!("[{}] => #{}", name, tpe_id));
-        }
-    }
-
-    fn variant_name(&self) -> String {
-        "struct".to_string()
-    }
-}
-
-impl TypeData for TypeFunc {
-    fn get_display_params_into(&self, params: &mut Vec<String>) {
-        params.push(format!("#{} => #{}", self.inp, self.out));
-    }
-
-    fn variant_name(&self) -> String {
-        "func".to_string()
-    }
-}
-
-impl TypeData for TypePolicy {
-    fn get_display_params_into(&self, params: &mut Vec<String>) {
-        params.push(format!(
-            "policy='[{}]'",
-            self.chain
-                .iter()
-                .map(|p| match p {
-                    PolicySpec::Simple(pol_id) => format!(
-                        "'{}'",
-                        with_store(|s| s.get_policy(*pol_id).unwrap().name.clone())
-                    ),
-                    PolicySpec::PerEffect(p) => with_store(|s| format!(
-                        "{{create='{}', update='{}', delete='{}', none='{}'}}",
-                        p.create
-                            .map(|pol_id| s.get_policy(pol_id).unwrap().name.as_str())
-                            .unwrap_or("null"),
-                        p.update
-                            .map(|pol_id| s.get_policy(pol_id).unwrap().name.as_str())
-                            .unwrap_or("null"),
-                        p.delete
-                            .map(|pol_id| s.get_policy(pol_id).unwrap().name.as_str())
-                            .unwrap_or("null"),
-                        p.none
-                            .map(|pol_id| s.get_policy(pol_id).unwrap().name.as_str())
-                            .unwrap_or("null"),
-                    )),
-                })
-                .collect::<Vec<_>>()
-                .join(", ")
-        ));
-    }
-
-    fn variant_name(&self) -> String {
-        "policy".to_string()
-    }
-}
-
-impl WrapperTypeData for TypePolicy {
-    fn get_wrapped_type<'a>(&self, store: &'a Store) -> Option<&'a Type> {
-        store.get_type(self.tpe).ok()
     }
 }
